+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# next
+
+- Add hooks, `Lua` actions that run before or after Lua programs. To upgrade
+  from v0.1, just add `defaultHooks` as the first argument to `check{,'}`.
+
 # [0.1.1.0] -- 2025-04-22
 
 [0.1.1.0]: https://github.com/GaloisInc/oughta/releases/tag/v0.1.1.0